+++ CHANGES.rst
@@ -3,10 +3,14 @@ Recent Changes
 
 For earlier changes, see the `HISTORY.rst <HISTORY.rst>`_.
 
-3.0.1 (unreleased)
+3.1.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Python 3 support
+
+- Minor convenience change to API: When instantiating primary or
+  secondary servers, you can pass a file-name instead of a storage
+  instance and a file storage will be created automatically.
 
 
 3.0.0 (2017-04-04)